requirements.txt
pygame-ce
pyscreenshot
pillow
pywin32
